src/renderer/state/job/test/selectors.test.ts
import { expect } from "chai";

import {
  JSSJob,
  JSSJobStatus,
  UploadStage,
} from "../../../services/job-status-client/types";
import {
  mockFailedUploadJob,
  mockState,
  mockSuccessfulUploadJob,
  mockWorkingUploadJob,
  nonEmptyJobStateBranch,
} from "../../test/mocks";
import {
  getIsSafeToExit,
  getJobIdToUploadJobMap,
  getJobsForTable,
} from "../selectors";

describe("Job selectors", () => {
  describe("getJobsForTable", () => {
    it("converts jobs in redux store to objects used by upload summary page", () => {
      const jobs = [...nonEmptyJobStateBranch.uploadJobs];
      const jobTableRows = getJobsForTable({
        ...mockState,
        job: {
          ...nonEmptyJobStateBranch,
          copyProgress: {
            [mockWorkingUploadJob.jobId]: {
              completedBytes: 2,
              totalBytes: 100,
            },
          },
        },
      });

      expect(jobTableRows.length).to.equal(jobs.length);
      let foundWorkingJob = false;
      jobTableRows.forEach((jobTableRow) => {
        const match = jobs.find((job) => {
          return (
            job.jobName === jobTableRow.jobName &&
            job.jobId === jobTableRow.jobId &&
            job.currentStage === jobTableRow.currentStage &&
            job.status === jobTableRow.status
          );
        });
        if (jobTableRow.status === JSSJobStatus.WORKING) {
          expect(jobTableRow.progress).to.not.be.undefined;
          foundWorkingJob = true;
        }
        expect(match).to.not.be.undefined;
      });
      expect(foundWorkingJob).to.be.true;
    });

    it("hides any jobs that are duplicates of the original", () => {
      // Arrange
      const mockReplacedJob1: JSSJob = {
        ...mockFailedUploadJob,
        created: new Date("Oct 2, 2020 03:24:00"),
        jobId: "replacement1",
      };
      const mockReplacedJob2: JSSJob = {
        ...mockFailedUploadJob,
        created: new Date("Oct 3, 2020 03:24:00"),
        jobId: "replacement2",
        status: JSSJobStatus.RETRYING,
        serviceFields: {
          ...mockFailedUploadJob.serviceFields,
          originalJobId: mockReplacedJob1.jobId,
        },
      };
      const expectedJob: JSSJob = {
        ...mockFailedUploadJob,
        created: new Date("Oct 1, 2020 03:24:00"),
        serviceFields: {
          files: [],
          lastModified: {},
          md5: {},
          originalJobId: mockReplacedJob2.jobId,
          type: "upload",
          uploadDirectory: "/foo",
        },
      };

      // Act
      const rows = getJobsForTable({
        ...mockState,
        job: {
          ...mockState.job,
          uploadJobs: [expectedJob, mockReplacedJob1, mockReplacedJob2],
        },
      });

      // Assert
      expect(rows).to.be.lengthOf(1);
      expect(rows[0].jobId).to.equal(expectedJob.jobId);
    });
  });

  describe("getIsSafeToExit", () => {
    it("returns false if an upload job is in progress and in client upload stage", () => {
      const isSafeToExit = getIsSafeToExit({
        ...mockState,
        job: {
          ...mockState.job,
          uploadJobs: [
            {
              ...mockWorkingUploadJob,
              currentStage: UploadStage.WAITING_FOR_CLIENT_COPY,
            },
          ],
        },
      });
      expect(isSafeToExit).to.be.false;
    });

    it("returns true if no jobs", () => {
      const isSafeToExit = getIsSafeToExit(mockState);
      expect(isSafeToExit).to.be.true;
    });

    it("returns true if an upload job is in progress and not in client upload stage", () => {
      const isSafeToExit = getIsSafeToExit({
        ...mockState,
        job: {
          ...mockState.job,
          uploadJobs: [
            {
              ...mockWorkingUploadJob,
              currentStage: UploadStage.COMPLETE,
            },
          ],
        },
      });
      expect(isSafeToExit).to.be.true;
    });

    it("returns true if there are no in progress jobs", () => {
      const isSafeToExit = getIsSafeToExit({
        ...mockState,
        job: {
          ...mockState.job,
          uploadJobs: [mockFailedUploadJob, mockSuccessfulUploadJob],
        },
      });
      expect(isSafeToExit).to.be.true;
    });
  });

  describe("getJobIdToUploadJobMap", () => {
    it("converts a list of jobs to a map of jobId's to jobs", () => {
      const map = getJobIdToUploadJobMap({
        ...mockState,
        job: {
          ...mockState.job,
          uploadJobs: [mockWorkingUploadJob, mockSuccessfulUploadJob],
        },
      });
      expect(map.size).to.equal(2);
      expect(map.get(mockWorkingUploadJob.jobId)).to.equal(
        mockWorkingUploadJob
      );
      expect(map.get(mockSuccessfulUploadJob.jobId)).to.equal(
        mockSuccessfulUploadJob
      );
    });
  });
});
